+++ sdk/recoveryservices/arm-recoveryservices/src/models/index.ts
@@ -296,6 +296,13 @@ export interface VaultProperties {
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly moveState?: ResourceMoveState;
+  /**
+   * Backup storage version
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly backupStorageVersion?: BackupStorageVersion;
+  /** The redundancy settings of the resource. */
+  redundancySettings?: VaultPropertiesRedundancySettings;
 }
 
 /** Details for upgrading vault. */
@@ -474,6 +481,20 @@ export interface VaultPropertiesMoveDetails {
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly targetResourceId?: string;
+}
+
+/** The redundancy settings of the resource. */
+export interface VaultPropertiesRedundancySettings {
+  /**
+   * The storage redundancy setting for the Vault
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly standardTierStorageRedundancy?: StandardTierStorageRedundancy;
+  /**
+   * Showing whether cross region restore is enabled or not
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly crossRegionRestore?: CrossRegionRestore;
 }
 
 /** Identifies the unique system identifier for each Azure resource. */
@@ -891,6 +912,58 @@ export enum KnownResourceMoveState {
  * **PartialSuccess**
  */
 export type ResourceMoveState = string;
+
+/** Known values of {@link BackupStorageVersion} that the service accepts. */
+export enum KnownBackupStorageVersion {
+  V1 = "V1",
+  V2 = "V2",
+  Unassigned = "Unassigned"
+}
+
+/**
+ * Defines values for BackupStorageVersion. \
+ * {@link KnownBackupStorageVersion} can be used interchangeably with BackupStorageVersion,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **V1** \
+ * **V2** \
+ * **Unassigned**
+ */
+export type BackupStorageVersion = string;
+
+/** Known values of {@link StandardTierStorageRedundancy} that the service accepts. */
+export enum KnownStandardTierStorageRedundancy {
+  GeoRedundant = "GeoRedundant",
+  LocallyRedundant = "LocallyRedundant",
+  ZoneRedundant = "ZoneRedundant"
+}
+
+/**
+ * Defines values for StandardTierStorageRedundancy. \
+ * {@link KnownStandardTierStorageRedundancy} can be used interchangeably with StandardTierStorageRedundancy,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **GeoRedundant** \
+ * **LocallyRedundant** \
+ * **ZoneRedundant**
+ */
+export type StandardTierStorageRedundancy = string;
+
+/** Known values of {@link CrossRegionRestore} that the service accepts. */
+export enum KnownCrossRegionRestore {
+  Enabled = "Enabled",
+  Disabled = "Disabled"
+}
+
+/**
+ * Defines values for CrossRegionRestore. \
+ * {@link KnownCrossRegionRestore} can be used interchangeably with CrossRegionRestore,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **Enabled** \
+ * **Disabled**
+ */
+export type CrossRegionRestore = string;
 
 /** Known values of {@link SkuName} that the service accepts. */
 export enum KnownSkuName {