+++ sdk/recoveryservices/arm-recoveryservices/src/models/parameters.ts
@@ -74,7 +74,7 @@ export const subscriptionId: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2021-08-01",
+    defaultValue: "2022-03-15-preview",
     isConstant: true,
     serializedName: "api-version",
     type: {