+++ sdk/recoveryservices/arm-recoveryservices/src/operations/vaults.ts
@@ -253,12 +253,10 @@ export class VaultsImpl implements Vaults {
       { resourceGroupName, vaultName, vault, options },
       createOrUpdateOperationSpec
     );
-    const poller = new LroEngine(lro, {
+    return new LroEngine(lro, {
       resumeFrom: options?.resumeFrom,
       intervalInMs: options?.updateIntervalInMs
     });
-    await poller.poll();
-    return poller;
   }
 
   /**
@@ -362,12 +360,10 @@ export class VaultsImpl implements Vaults {
       { resourceGroupName, vaultName, vault, options },
       updateOperationSpec
     );
-    const poller = new LroEngine(lro, {
+    return new LroEngine(lro, {
       resumeFrom: options?.resumeFrom,
       intervalInMs: options?.updateIntervalInMs
     });
-    await poller.poll();
-    return poller;
   }
 
   /**